Combine Overview with MicroserviceViewScreen
